requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile '.\requirements.in'
#
aiofiles==23.2.1
    # via aiogram
aiogram==3.4.1
    # via -r .\requirements.in
aiohttp==3.9.5
    # via aiogram
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via aiohttp
certifi==2025.7.14
    # via aiogram
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.3
    # via sqlalchemy
idna==3.10
    # via yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
propcache==0.3.2
    # via yarl
pydantic==2.5.3
    # via aiogram
pydantic-core==2.14.6
    # via pydantic
python-dotenv==1.0.1
    # via -r .\requirements.in
sqlalchemy==2.0.41
    # via -r .\requirements.in
typing-extensions==4.14.1
    # via
    #   aiogram
    #   aiosignal
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
yarl==1.20.1
    # via aiohttp
